# internal/api/openapi.yaml
openapi: 3.1.0
info:
  title: AMDL Backend API
  version: 1.0.0
  description: |
    Apple Music 下载任务与 wrapper-manager 管理接口。

    多用户模式（auth.enabled 为 true）下，后端信任反向代理注入的 X-User / X-Email 头解析身份；
    未注册或被禁用的用户返回 403。auth.enabled 为 false 时所有请求视为内置管理员。
    Wrapper 与 Users 分组接口仅管理员可用；普通用户只能访问自己的下载任务，
    访问他人任务一律返回 404。
servers:
  - url: /
...
paths:
...
  /api/v1/me:
    get:
      tags: [Users]
      summary: 查询当前登录用户
      description: 返回当前身份的用户名、角色和头像，供前端渲染。单用户模式返回内置管理员。
      operationId: getMe
      responses:
        '200':
          description: 当前用户信息。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MeResponse'
        '403':
          $ref: '#/components/responses/Forbidden'
  /api/v1/users:
    get:
      tags: [Users]
      summary: 列出用户（仅管理员）
      operationId: listUsers
      responses:
        '200':
          description: 用户列表。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
    post:
      tags: [Users]
      summary: 创建用户（仅管理员）
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
      responses:
        '201':
          description: 用户已创建。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: 用户名、别名或邮箱已被占用。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/users/{user_id}:
    get:
      tags: [Users]
      summary: 查询用户（仅管理员）
      operationId: getUser
      parameters:
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: 用户详情。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [Users]
      summary: 更新用户（仅管理员）
      description: 部分更新，仅提交的字段生效；aliases / emails 提交时整组替换。
      operationId: updateUser
      parameters:
        - $ref: '#/components/parameters/UserID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserRequest'
      responses:
        '200':
          description: 更新后的用户。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: 别名或邮箱已被占用。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      tags: [Users]
      summary: 禁用用户（仅管理员）
      description: 软禁用（enabled=false），保留任务归属供审计；不物理删除。
      operationId: disableUser
      parameters:
        - $ref: '#/components/parameters/UserID'
      responses:
        '200':
          description: 用户已禁用。
          content:
            application/json:
              schema:
                type: object
                required: [status, id]
                properties:
                  status:
                    type: string
                    const: disabled
                  id:
                    type: string
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /api/v1/wrapper/status:
    get:
      tags: [Wrapper]
      summary: 查询 wrapper-manager 状态（仅管理员）
      description: 探测 wrapper-manager 连通性与登录/区域等运行状态；下载前或账号管理前使用。
      operationId: getWrapperStatus
      responses:
        '200':
          description: Wrapper 状态。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WrapperStatus'
        '403':
          $ref: '#/components/responses/Forbidden'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/wrapper/login:
    post:
      tags: [Wrapper]
      summary: 登录 wrapper 账号
      description: 登录可能直接成功，也可能返回 login_id 要求两步验证码。
      operationId: loginWrapper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: 登录成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggedInResponse'
        '202':
          description: 需要提交两步验证码。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NeedsTwoStepResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /api/v1/wrapper/login/{login_id}/2fa:
    post:
      tags: [Wrapper]
      summary: 提交两步验证码
      operationId: submitWrapperTwoStepCode
      parameters:
        - $ref: '#/components/parameters/LoginID'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TwoStepRequest'
      responses:
        '200':
          description: 登录成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggedInResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '502':
          $ref: '#/components/responses/BadGateway'
        '504':
          $ref: '#/components/responses/GatewayTimeout'
  /api/v1/wrapper/logout:
    post:
      tags: [Wrapper]
      summary: 登出 wrapper 账号
      operationId: logoutWrapper
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LogoutRequest'
      responses:
        '200':
          description: 登出成功。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoggedOutResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '502':
          $ref: '#/components/responses/BadGateway'
  /api/v1/quality:
    post:
      tags: [Quality]
      summary: 查询歌曲可用音质
      description: 解析 Apple Music 单曲 URL 或带 i 参数的专辑 URL，返回可用 codec、码率、采样率和位深摘要；不会创建下载任务。
      operationId: queryQuality
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/QualityRequest'
      responses:
        '200':
          description: 可用音质摘要。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '502':
          $ref: '#/components/responses/BadGateway'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/downloads:
    get:
      tags: [Downloads]
      summary: 列出下载任务
      operationId: listDownloads
      description: 普通用户仅返回自己的任务；管理员返回全部任务，可用 user 参数按用户名过滤。
      parameters:
        - name: limit
          in: query
          description: 返回数量，默认 50，最大 200。
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: user
          in: query
          description: 仅管理员有效：按用户名过滤；未知用户名返回空列表。
          schema:
            type: string
      responses:
        '200':
          description: 下载任务列表。
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Job'
        '500':
          $ref: '#/components/responses/InternalError'
    post:
      tags: [Downloads]
      summary: 创建下载任务
      operationId: createDownload
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DownloadRequest'
      responses:
        '202':
          description: 任务已入队。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Job'
        '400':
          $ref: '#/components/responses/BadRequest'
        '422':
          description: URL 对应的内容或 storefront 不受支持。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          $ref: '#/components/responses/ServiceUnavailable'
  /api/v1/downloads/{job_id}:
    get:
      tags: [Downloads]
      summary: 查询下载任务详情
      operationId: getDownload
      parameters:
        - $ref: '#/components/parameters/JobID'
      responses:
        '200':
          description: 任务、任务项与重试策略。
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DownloadDetail'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  parameters:
    LoginID:
      name: login_id
      in: path
      required: true
      schema:
        type: string
      example: 4f0f2c79acda4c28b69ebf4f702f8a51
    JobID:
      name: job_id
      in: path
      required: true
      schema:
        type: string
      example: job_01JZ0000000000000000000000
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        type: string
      example: user_01JZ0000000000000000000000
  responses:
    BadRequest:
      description: 请求参数错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: 登录或验证码错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: 身份未注册、已禁用或权限不足。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: 资源不存在或登录会话已过期。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Conflict:
      description: 账号已经登录或会话正在验证。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    BadGateway:
      description: wrapper-manager 返回异常。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ServiceUnavailable:
      description: 服务暂时不可用。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    GatewayTimeout:
      description: wrapper 登录超时。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalError:
      description: 服务器内部错误。
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
        storefront:
          type: string
        supported_storefronts:
          type: array
          items:
            type: string
    WrapperStatus:
      type: object
      required: [ready, status, regions, client_count, accounts_supported]
      properties:
        ready:
          type: boolean
        status:
          type: boolean
        regions:
          type: array
          items:
            type: string
        client_count:
          type: integer
          format: int32
        accounts_supported:
          type: boolean
          description: Whether wrapper-manager returned the optional accounts field; false means the connected manager version does not support it.
        accounts:
          type: array
          description: Logged-in wrapper account identifiers when supported by wrapper-manager.
          items:
            type: string
...
    RetryPolicy:
      type: object
      required: [operation_retries, first_codec_retries, fallback_codec_retries]
      properties:
        operation_retries:
          type: integer
        first_codec_retries:
          type: integer
        fallback_codec_retries:
          type: integer
...
    LoginRequest:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          format: password
          writeOnly: true
          example: example-password
    TwoStepRequest:
      type: object
      required: [two_step_code]
      properties:
        two_step_code:
          type: string
          example: '123456'
    LogoutRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
          example: user@example.com
    LoggedInResponse:
      type: object
      required: [status]
      properties:
        status:
          type: string
          const: logged_in
    NeedsTwoStepResponse:
      type: object
      required: [status, login_id]
      properties:
        status:
          type: string
          const: needs_2fa
        login_id:
          type: string
    LoggedOutResponse:
      type: object
      required: [status, username]
      properties:
        status:
          type: string
          const: logged_out
        username:
          type: string
    DownloadRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: https://music.apple.com/cn/album/example/123456789?i=987654321
        force:
          type: boolean
          default: false
    QualityRequest:
      type: object
      required: [url]
      properties:
        url:
          type: string
          format: uri
          example: https://music.apple.com/cn/album/one-last-kiss/1542953969?i=1542953977
    QualitySong:
      type: object
      required: [id, name, artist, album]
      properties:
        id: {type: string}
        name: {type: string}
        artist: {type: string}
        album: {type: string}
    QualityOption:
      type: object
      required: [id, label, available]
      properties:
        id:
          type: string
          enum: [aac, aac-binaural, aac-downmix, alac, ec3, ac3]
        label: {type: string}
        available: {type: boolean}
        codec: {type: string}
        codec_id: {type: string}
        channels: {type: integer}
        bit_depth: {type: integer}
        sample_rate:
          type: integer
          description: Hz.
        bitrate:
          type: integer
          description: bits per second.
        description: {type: string}
    QualityResponse:
      type: object
      required: [input, storefront, type, adam_id, song, qualities]
      properties:
        input: {type: string}
        storefront: {type: string}
        type:
          type: string
          const: song
        adam_id: {type: string}
        song:
          $ref: '#/components/schemas/QualitySong'
        qualities:
          type: array
          items:
            $ref: '#/components/schemas/QualityOption'
    MeResponse:
      type: object
      required: [user_id, username, role, avatar_url]
      properties:
        user_id:
          type: string
          description: 单用户模式下为空字符串。
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        avatar_url:
          type: string
    User:
      type: object
      required: [id, username, role, avatar_url, enabled, aliases, emails, created_at, updated_at]
      properties:
        id: {type: string}
        username:
          type: string
          description: 须匹配 ^[a-z0-9_-]{1,32}$，同时用作用户下载目录名。
        role:
          type: string
          enum: [admin, user]
        avatar_url: {type: string}
        enabled: {type: boolean}
        aliases:
          type: array
          items: {type: string}
        emails:
          type: array
          items: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    CreateUserRequest:
      type: object
      required: [username]
      properties:
        username:
          type: string
          description: 须匹配 ^[a-z0-9_-]{1,32}$。
        role:
          type: string
          enum: [admin, user]
          default: user
        avatar_url: {type: string}
        aliases:
          type: array
          items: {type: string}
        emails:
          type: array
          items: {type: string}
    UpdateUserRequest:
      type: object
      description: 所有字段可选，仅提交的字段生效。
      properties:
        role:
          type: string
          enum: [admin, user]
        enabled: {type: boolean}
        avatar_url: {type: string}
        aliases:
          type: array
          items: {type: string}
        emails:
          type: array
          items: {type: string}
    Job:
      type: object
      required: [id, input, type, force, status, total_items, done_items, failed_items, created_at, updated_at]
      properties:
        id:
          type: string
        user_id:
          type: string
          description: 任务归属用户 ID；历史或单用户任务为空。
        username:
          type: string
          description: 归属用户名，便于展示。
        input:
          type: string
        type:
          type: string
          enum: [song, album, playlist]
        storefront:
          type: string
        force:
          type: boolean
        status:
          type: string
          enum: [queued, running, completed, failed, cancelled]
        total_items:
          type: integer
        done_items:
          type: integer
        failed_items:
          type: integer
        error:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    JobItem:
      type: object
      required: [id, job_id, adam_id, kind, index, status, progress, created_at, updated_at]
      properties:
        id: {type: string}
        job_id: {type: string}
        adam_id: {type: string}
        kind: {type: string}
        index: {type: integer}
        title: {type: string}
        artist: {type: string}
        album: {type: string}
        status:
          type: string
          enum: [queued, resolving, downloading, decrypting, remuxing, tagging, saving, completed, failed, skipped_existing, cancelled]
        progress: {type: number, format: double}
        codec: {type: string}
        retry_kind: {type: string}
        attempt: {type: integer}
        max_attempts: {type: integer}
        status_message: {type: string}
        output_path: {type: string}
        error: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
    DownloadDetail:
      type: object
      required: [job, items, retry_policy]
      properties:
        job:
          $ref: '#/components/schemas/Job'
        items:
          type: array
          items:
            $ref: '#/components/schemas/JobItem'
        retry_policy:
          $ref: '#/components/schemas/RetryPolicy'